nemo-backend/requirements.txt
fastapi
uvicorn
google-generativeai
google-genai
pydantic